+++ tests/unit_tests/test_conf.c
@@ -65,24 +65,29 @@ void test_conf(void **state)
 	// Test that we reject this json log file location
 	char cli_json_log_file_location_wrong[] =
 		"tests/unit_tests/test_sentrypeer_json.log";
-	assert_int_equal(set_json_log_file_location(config,
-						    cli_json_log_file_location_wrong),
+	assert_int_equal(set_json_log_file_location(
+				 config, cli_json_log_file_location_wrong),
 			 EXIT_FAILURE);
 
 	// Test that we accept a valid json log file location
-	char cli_json_log_file_location[] = "/tests/unit_tests/test_sentrypeer_json.log";
-	assert_int_equal(set_json_log_file_location(config, cli_json_log_file_location),
+	char cli_json_log_file_location[] =
+		"/tests/unit_tests/test_sentrypeer_json.log";
+	assert_int_equal(set_json_log_file_location(config,
+						    cli_json_log_file_location),
 			 EXIT_SUCCESS);
 
-	assert_int_equal(setenv("SENTRYPEER_JSON_LOG_FILE", cli_json_log_file_location, 1),
+	assert_int_equal(setenv("SENTRYPEER_JSON_LOG_FILE",
+				cli_json_log_file_location, 1),
 			 EXIT_SUCCESS);
 	// Test that we set our json log file location via the environment variable SENTRYPEER_JSON_LOG_FILE
-	assert_int_equal(set_json_log_file_location(config, NULL), EXIT_SUCCESS);
+	assert_int_equal(set_json_log_file_location(config, NULL),
+			 EXIT_SUCCESS);
 	assert_string_equal(config->json_log_file, cli_json_log_file_location);
 
 	assert_int_equal(unsetenv("SENTRYPEER_JSON_LOG_FILE"), EXIT_SUCCESS);
 	// Test that we set our own json log file location to cwd of runner
-	assert_int_equal(set_json_log_file_location(config, NULL), EXIT_SUCCESS);
+	assert_int_equal(set_json_log_file_location(config, NULL),
+			 EXIT_SUCCESS);
 
 	// Test non db file env vars
 	assert_int_equal(setenv("SENTRYPEER_DEBUG", "1", 1), EXIT_SUCCESS);
@@ -103,15 +108,11 @@ void test_conf(void **state)
 	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
 	assert_false(config->sip_mode);
 
-
-	assert_int_equal(setenv("SENTRYPEER_API", "1", 1),
-			 EXIT_SUCCESS);
+	assert_int_equal(setenv("SENTRYPEER_API", "1", 1), EXIT_SUCCESS);
 	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
 	assert_true(config->api_mode);
 
-
-	assert_int_equal(setenv("SENTRYPEER_WEB_GUI", "1", 1),
-			 EXIT_SUCCESS);
+	assert_int_equal(setenv("SENTRYPEER_WEB_GUI", "1", 1), EXIT_SUCCESS);
 	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
 	assert_true(config->web_gui_mode);
 
@@ -123,10 +124,22 @@ void test_conf(void **state)
 	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
 	assert_true(config->json_log_mode);
 
-	assert_int_equal(setenv("SENTRYPEER_PEER_TO_PEER", "1", 1), EXIT_SUCCESS);
+	assert_int_equal(setenv("SENTRYPEER_PEER_TO_PEER", "1", 1),
+			 EXIT_SUCCESS);
 	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
-	assert_true(config->p2p_lan_mode);
 	assert_true(config->p2p_dht_mode);
+
+	// Check default bootstrap node is set
+	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
+	assert_string_equal(config->p2p_bootstrap_node,
+			    SENTRYPEER_BOOTSTRAP_NODE);
+
+	// Set our own node
+	char bootstrap_node[] = "bootstrap.example.com";
+	assert_int_equal(setenv("SENTRYPEER_BOOTSTRAP_NODE", bootstrap_node, 1),
+			 EXIT_SUCCESS);
+	assert_int_equal(process_env_vars(config), EXIT_SUCCESS);
+	assert_string_equal(config->p2p_bootstrap_node, bootstrap_node);
 
 	sentrypeer_config_destroy(&config);
 	assert_null(config);