+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- New `-b` flag to set your own DHT bootstrap node (default: `bootstrap.sentrypeer.org:4222`) 
 
 ## [1.4.0] - 2022-03-29
 - Peer to Peer mode with CLI (`-p`) and ENV options to enable, added for DHT node using the OpenDHT library